feat: emulate the Hyperopt interface
